+++ contentful_management/snapshot.py
@@ -1,4 +1,4 @@
-from .resource import FieldsResource
+from .resource import Resource
 
 
 """
@@ -14,13 +14,16 @@ API Reference: https://www.contentful.com/developers/docs/references/content-del
 """
 
 
-class Snapshot(FieldsResource):
+class Snapshot(Resource):
     """
     API Reference: https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/content-types
     """
 
     def __init__(self, item, **kwargs):
         super(Snapshot, self).__init__(item, **kwargs)
+
+        from .entry import Entry
+        self.snapshot = Entry(item['snapshot'], **kwargs)
 
     @classmethod
     def base_url(klass, space_id, entry_id, resource_id=None):
@@ -31,6 +34,15 @@ class Snapshot(FieldsResource):
             entry_id,
             resource_id if resource_id is not None else ''
         )
+
+    def to_json(self):
+        """Returns the JSON Representation of the Resource"""
+
+        result = super(Snapshot, self).to_json()
+        result.update({
+            'snapshot': self.snapshot.to_json(),
+        })
+        return result
 
     def __repr__(self):
         return "<Snapshot id='{0}'>".format(